requirements.txt
flask
pillow
requests
pandas
scikit-learn
torch
torchvision
opencv-python
